Corrected behavior of fallback in get()
It seems that fallback was introduced in 781165db. The intended
effect of this commit is

    webL10n.get('non_existent', null, 'fallback') === 'fallback'

However, the implementation was broken, because the fallback was
assigned but not used, see
https://github.com/fabi1cazenave/webL10n/blob/781165db/l10n.js#L206-L211

A later update,
https://github.com/fabi1cazenave/webL10n/commit/d798e43f#L1L294
(accidentally?) changed the API to something useless. In order to get
the fallback to work, one has to use:

   webL10n.get('non_existent', null, {textContent: 'fallback'});
where "textContent" is the value of the private variable gTextProp.
This API makes no sense, so I have updated the API to behave as
originally intended:

   webL10n.get('non_existent', null, 'fallback') === 'fallback';